Remove XPU code; remove cpu example; add UT
